Various changes to benchmarks to assist with PicoC parsing
No functionality should have changed, however none of these benchmarks
currently run through PicoC for various reasons (either returning
incorrect values, using unsupported features, causing parsing issues, or
causing segfaults).

These changes are potentially progress towards getting these benchmarks
running eventually, however PicoC changes will also be reqired for that
to happen.

+++ bench/sequential/ammunition/bits.c
@@ -239,7 +239,7 @@ ammunition_bit_string_move ( void *to, int to_bit_displacement,
   to_bit_displacement %= CHAR_BIT;
   current_from_byte += from_bit_displacement / CHAR_BIT;
   from_bit_displacement %= CHAR_BIT;
-  if ( current_from_byte > current_to_byte
+  if ( (long)current_from_byte > (long)current_to_byte
        || ( current_from_byte == current_to_byte
             && from_bit_displacement > to_bit_displacement ) )
     ammunition_bit_string_copy ( current_to_byte, to_bit_displacement,
